% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_group.R
\name{common_group}
\alias{common_group}
\title{overview of data frame with group}
\usage{
common_group(dat, group_vec, ref_vec)
}
\arguments{
\item{dat}{a dataframe, preferably the wide format}

\item{group_vec}{grouping vector, contain column names to be examine}

\item{ref_vec}{reference vector, contain column names of target to be check}
}
\value{
a data table with range/elements of each column
}
\description{
\code{common_group} find the common elements of selected groups
}
\examples{
common_group(mtcars,"cyl","gear")
common_group(iris,"Species","Sepal.Length")
}
\author{
Tien-Cheng Wang
}
\keyword{dataframe}
\keyword{groups}
\keyword{in}
\keyword{overview}
