% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_ue.R
\name{df_ue}
\alias{df_ue}
\title{Unique Elements}
\usage{
df_ue(df, coln, pattern = NULL)
}
\arguments{
\item{df}{a dataframe, include columns xvar and yvar}

\item{coln}{unquote column name, if `coln`="cnames", then return colnames}

\item{pattern}{`NULL`, else match pattern in unique coln values}
}
\value{
unique values of a column
}
\description{
\code{df_ue} find unique elements plus matching services
}
\examples{
  library(dplyr)
  df_ue(iris,Species)
  df_ue(iris,Species,"virg")
}
\author{
Tien-Cheng Wang
}
\keyword{match}
\keyword{pattern}
\keyword{unique}
