% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_readme.R
\name{parallel_readme}
\alias{parallel_readme}
\title{readme script for parallel computing}
\usage{
parallel_readme()
}
\value{
scripts for initialization
}
\description{
\code{parallel_readme} provide framework for parallel
}
\examples{
parallel_readme()

}
\author{
Tien-Cheng Wang
}
\keyword{parallel}
