% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_121.R
\name{plot_121}
\alias{plot_121}
\title{plot 1 to 1 line to examine linearity}
\usage{
plot_121(df, xvar, yvar, color = NULL, labx = NULL, laby = NULL, ...)
}
\arguments{
\item{df}{a dataframe, include columns xvar and yvar}

\item{xvar}{unquoted variable for colmun name for x axis, should be categorical}

\item{yvar}{unquoted variable for column name for y axis, should be numeric}

\item{color}{`NULL`, unquoted variable for colmun name for group}

\item{labx}{string for x axis title, default is xvar.}

\item{laby}{string for y axis title, default is yvar}
}
\value{
a ggplot with 1:1 line
}
\description{
\code{plot_121} plot 1 to 1 line to examine linearity
}
\examples{
  pacman::p_load(ggplot2,toolPhD,ggpmisc)
  plot_121(iris,Sepal.Length,Petal.Length,color=Species)
  plot_121(iris,Sepal.Length,Petal.Length)
}
\author{
Tien-Cheng Wang
}
\keyword{violoin}
