% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_pvalue.R
\name{sig_pvalue}
\alias{sig_pvalue}
\title{tidy digit of numbers}
\usage{
sig_pvalue(vec)
}
\arguments{
\item{vec}{a vector, which may include different scale of numeric number.}
}
\value{
a character vector, which turn pvalue into significance for each element
}
\description{
\code{sig_pvalue} display the tidy unit of your observations
}
\examples{
vec <- c(1.1,0.1,0.01,0.001,0.0001,0.00001)
sig_pvalue(vec)
}
\author{
Tien-Cheng Wang
}
\keyword{display}
\keyword{for}
\keyword{format}
\keyword{scientific}
