% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_cor.R
\name{sma_cor}
\alias{sma_cor}
\title{sma correlation matrix}
\usage{
sma_cor(mat)
}
\arguments{
\item{dat}{a dataframe, include only numeric columns for correlation}
}
\value{
a list of three elements: 1. `r.mat`r matrix, 2.`p.mat` p-value matrix and 3 `s.mat` slope matrix.
}
\description{
\code{sma_cor} calculate sma correlation matrix for each complete pairs.
}
\examples{
# numeric dataframe
sma_cor(mtcars[,1:7])
}
\author{
Tien-Cheng Wang
}
\keyword{axis}
\keyword{major}
\keyword{standardised}
