% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_plot.R
\name{sma_plot}
\alias{sma_plot}
\title{sma plot}
\usage{
sma_plot(datf, xvar, yvar)
}
\arguments{
\item{datf}{a dataframe, include columns x and y}

\item{xvar}{unquoted string for column name x}

\item{yvar}{unquoted string for column name y}
}
\value{
a ggplot with regression line and r2 text
}
\description{
\code{sma_plot} plot sma correlation analysis add regression line and add r2
}
\examples{
sma_plot(mtcars,xvar=mpg,yvar = cyl)
}
\author{
Tien-Cheng Wang
}
\keyword{axis}
\keyword{major}
\keyword{standardised}
