% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_phd_facet.R
\name{theme_phd_facet}
\alias{theme_phd_facet}
\title{theme phd facet}
\usage{
theme_phd_facet(
  ax.txt.siz = NULL,
  ax.tit.siz = NULL,
  lgd.txt.siz = NULL,
  lgd.tit.siz = NULL,
  strp.txt.siz = NULL,
  t = NULL,
  r = NULL,
  b = NULL,
  l = NULL,
  ...
)
}
\arguments{
\item{ax.txt.siz}{axis text size}

\item{ax.tit.siz}{axis title size}

\item{lgd.txt.siz}{legend text size}

\item{lgd.tit.siz}{legend title size}

\item{strp.txt.siz}{strip(facet) text size}

\item{t}{distance to top border}

\item{r}{distance to right border}

\item{b}{distance to bottom border}

\item{l}{distance to left border}
}
\value{
aesthetic facet frame
}
\description{
\code{theme_phd_facet} plot ggplot like a pro for facet
}
\examples{
library(ggplot2)
p <- ggplot(mtcars,aes(x=disp,y=hp)) + geom_point()+facet_grid(~cyl)
cowplot::plot_grid(p,p+theme_phd_facet())

}
\author{
Tien-Cheng Wang
}
\keyword{axis}
\keyword{major}
\keyword{standardised}
