% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_df.R
\name{view_df}
\alias{view_df}
\title{overview of data frame}
\usage{
view_df(x)
}
\arguments{
\item{x}{a dataframe, preferably the wide format}
}
\value{
a data table with range/elements of each column
}
\description{
\code{view_df} display the range/elements of each column
}
\examples{
vie_df(mtcars)
}
\author{
Tien-Cheng Wang
}
\keyword{dataframe}
\keyword{overview}
