% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_group.R
\name{view_group}
\alias{view_group}
\title{overview of data frame with group}
\usage{
view_group(dat, g_vec, t_vec = NULL)
}
\arguments{
\item{dat}{a dataframe, preferably the wide format}

\item{g_vec}{grouping vector, contain column names for grouping}

\item{t_vec}{target vector, contain column names for examining the range and elements.}
}
\value{
a data table with range/elements of each column
}
\description{
\code{view_group} display the range/elements of each column for selected group
}
\examples{
view_group(iris,c("Species"),c("Sepal.Length","Petal.Length"))
view_group(iris,c("Species"))
}
\author{
Tien-Cheng Wang
}
\keyword{dataframe}
\keyword{groups}
\keyword{in}
\keyword{overview}
