% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{violin_plot}
\alias{violin_plot}
\title{violin plot}
\usage{
violin_plot(df, xvar, yvar, orderx = F, labx = NULL, laby = NULL, ...)
}
\arguments{
\item{df}{a dataframe, include columns xvar and yvar}

\item{xvar}{unquoted variable for colmun name for x axis, should be categorical}

\item{yvar}{unquoted variable for column name for y axis, should be numeric}

\item{labx}{string for x axis title, default is xvar.}

\item{laby}{string for y axis title, default is yvar}

\item{oderx}{logical, default is FALSE, if TRUE, order the x axis based on y.}
}
\value{
a ggplot with violin plot
}
\description{
\code{violin_plot} plot distribution of numeric y over categorical x
}
\examples{
violin_plot(mtcars,cyl,mpg,orderx=T)
violin_plot(mtcars,cyl,mpg,labx="Number of cylinders",laby="Miles/gallon")
}
\author{
Tien-Cheng Wang
}
\keyword{violoin}
