% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_df.R
\name{cor_df}
\alias{cor_df}
\title{wrapper for correlation calculation}
\usage{
cor_df(dat, tar = NULL, method = "pearson")
}
\arguments{
\item{dat}{a dataframe, include *only* numeric columns for correlation}

\item{tar}{a dataframe, include *only* numeric columns for correlation}

\item{metohd}{default is NULL}
}
\value{
a list of three elements: 1. `r.mat`r matrix, 2.`p.mat` p-value matrix
}
\description{
\code{cor_df} calculate correlation matrix/stripe for each complete pairs.
}
\examples{
# numeric dataframe
cor_df(mtcars)
cor_df(mtcars,"cyl")

# dataframe contains character
cor_df(iris)
cor_df(iris[,-5])
cor_df(iris[,-5], method='sma')
cor_df(iris[,-5],'Sepal.Length' ,method='sma')

}
\author{
Tien-Cheng Wang
}
\keyword{correlation}
\keyword{pearson}
\keyword{sma}
\keyword{spearman}
