% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_plot.R
\name{sma_plot}
\alias{sma_plot}
\title{sma plot}
\usage{
sma_plot(data, x_string, y_string)
}
\arguments{
\item{data}{a dataframe, include columns x and y}

\item{x_string}{string for colmun name x}

\item{y_string}{string for column name y}
}
\value{
a ggplot with regression line and r2 text
}
\description{
\code{sma_plot} plot sma correlation analysis add regression line and add r2
}
\examples{
sma_plot(mtcars,x_string="mpg",y_string = "cyl")
}
\references{
\insertRef{warton2012}{toolPhD}
}
\author{
Tien-Cheng Wang
}
\keyword{axis}
\keyword{major}
\keyword{standardised}
